♻️ refactor:  URL起動状態をdomain層に移動
Issue #285
