+++ lib/domain/entities/url_launcher.freezed.dart
